refactor(io): align service filenames with nest conventions
