+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "eddie",
-    "version": "1.0.2",
+    "version": "1.0.3",
     "main": "dist/main.js",
     "bin": {
         "eddie": "dist/main.js"
@@ -8,7 +8,7 @@
     "scripts": {
         "build": "nest build",
         "lint": "eslint 'src/**/*.ts'",
-        "test": "vitest",
+        "test": "vitest --run",
         "dev": "nest start --watch"
     },
     "repository": {
