+++ src/cli/commands/context.command.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { ConfigService } from "../../config";
 import { ContextService } from "../../core/context";
 import { TokenizerService } from "../../core/tokenizers/strategy";
-import { LoggerService } from "../../io/logger";
+import { LoggerService } from "../../io";
 import type { CliArguments } from "../cli-arguments";
 import { CliOptionsService } from "../cli-options.service";
 import type { CliCommand } from "./cli-command";
