+++ src/core/engine/engine.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { ConfigModule } from "../../config/config.module";
 import { ContextModule } from "../context/context.module";
-import { IoModule } from "../../io/io.module";
+import { IoModule } from "../../io";
 import { EngineService } from "./engine.service";
 import { ProviderFactory } from "../providers";
 import { ToolRegistryFactory } from "../tools/registry";
