+++ src/io/io.module.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
-import { ConfirmService } from "./confirm";
-import { JsonlWriterService } from "./jsonl_writer";
-import { LoggerService } from "./logger";
-import { StreamRendererService } from "./stream_renderer";
+import { ConfirmService } from "./confirm.service";
+import { JsonlWriterService } from "./jsonl-writer.service";
+import { LoggerService } from "./logger.service";
+import { StreamRendererService } from "./stream-renderer.service";
 
 @Module({
   providers: [LoggerService, ConfirmService, JsonlWriterService, StreamRendererService],
