+++ src/io/stream-renderer.service.ts
