+++ test/integration/cli-runner.integration.test.ts
@@ -13,7 +13,7 @@ import { EngineService } from "../../src/core/engine";
 import { ConfigService } from "../../src/config";
 import { ContextService } from "../../src/core/context";
 import { TokenizerService } from "../../src/core/tokenizers/strategy";
-import { LoggerService } from "../../src/io/logger";
+import { LoggerService } from "../../src/io";
 import type { CliCommand } from "../../src/cli/commands/cli-command";
 import type { CliArguments } from "../../src/cli/cli-arguments";
 import type { EddieConfig } from "../../src/config/types";
